added a class with util functions for working with the orthography
the goal is to abstract some of the lower level details of working with
raw characters. this will make things like dealing with multiple
character letters (eg. 'ng') a whole lot easier. It should allow for
better suffixation algorithms since many of those work on the premiss of
syllables instead of raw characters.